refactor: shared.type 일부 이동
